+++ LibApophis/include/apophis/Utils/IImportReader.h
@@ -21,15 +21,15 @@ namespace Apophis { namespace Utils {
 		virtual std::unique_ptr<IImportReader> GetObject(const char* key) = 0;
 
 		virtual size_t Size() = 0;
-		virtual long long GetInt64(int index) = 0;
-		virtual real GetReal(int index) = 0;
-		virtual const char* GetString(int index) = 0;
-		virtual std::unique_ptr<IImportReader> GetObject(int index) = 0;
-		virtual std::unique_ptr<IImportReader> GetArray(int index) = 0;
+		virtual long long GetInt64(size_t index) = 0;
+		virtual real GetReal(size_t index) = 0;
+		virtual const char* GetString(size_t index) = 0;
+		virtual std::unique_ptr<IImportReader> GetObject(size_t index) = 0;
+		virtual std::unique_ptr<IImportReader> GetArray(size_t index) = 0;
 
 		// Helpers for common types
 		int GetInt32(const char* key) { return (int)GetInt64(key); }
-		int GetInt32(int index) { return (int)GetInt64(index); }
+		int GetInt32(size_t index) { return (int)GetInt64(index); }
 		size_t GetSize_t(const char* key) { return (size_t)GetInt64(key); }
 	};
 }}
